+++ lib/src/dsdiff/dsdiff_loader.dart
@@ -25,7 +25,7 @@ class DsdiffLoader extends ParserLoader {
       ..registerMapper('ID3v2.3', Id3v2TagMapper())
       ..registerMapper('ID3v2.4', Id3v2TagMapper());
 
-    final metadata = MetadataCollector(mapper);
+    final metadata = MetadataCollector(mapper, options);
     final parser = DsdiffParser(
       metadata: metadata,
       tokenizer: tokenizer,