+++ lib/src/mpeg/mpeg_loader.dart
@@ -38,7 +38,7 @@ class MpegLoader extends ParserLoader {
       ..registerMapper('ID3v2.3', Id3v2TagMapper())
       ..registerMapper('ID3v2.4', Id3v2TagMapper());
 
-    final metadata = MetadataCollector(mapper);
+    final metadata = MetadataCollector(mapper, options);
     metadata.setFormat(container: 'mp3');
 
     final parser = MpegParser(