+++ lib/src/musepack/musepack_loader.dart
@@ -32,7 +32,7 @@ class MusepackLoader extends ParserLoader {
       ..registerMapper('APEv2', Apev2TagMapper())
       ..registerMapper('matroska', MatroskaTagMapper());
 
-    final metadata = MetadataCollector(mapper);
+    final metadata = MetadataCollector(mapper, options);
     final parser = MusepackParser(
       metadata: metadata,
       tokenizer: tokenizer,