+++ lib/src/wav/wave_loader.dart
@@ -31,7 +31,7 @@ class WaveLoader extends ParserLoader {
       ..registerMapper('ID3v2.3', Id3v2TagMapper())
       ..registerMapper('ID3v2.4', Id3v2TagMapper());
 
-    final metadata = MetadataCollector(mapper);
+    final metadata = MetadataCollector(mapper, options);
 
     final parser = WaveParser(
       metadata: metadata,